requirements.txt
bs4
lxml
python-dotenv
tempfile